Fix Travis-CI Python 3.4 failure
Fix build failure on Travis-CI with Python 3.4 (v.3.5 passes) that appears to be caused by differences between versions in the handling of the builtin 'input' method when used with unittest.mock. The 'create=True' option added to all of the mock.patch.object instances makes the 'input' method available for the module under test.