Add baseline market ECM filter options
Add to the function that prompts the user for input to filter the list of active and inactive ECMs the ability to move ECMs from the active to the inactive list based on the user's specified preferences for a few of the applicable baseline market parameters. Parameters supported include climate zone, building sector (i.e., residential, commercial), and structure type. Add the corresponding tests, including required mocking and patching of user input.