Revise formatting of file names and paths
Add a custom CSS class to style file names and paths and set up global HTML span tags (in conf.py) available through substitutions to support applying the style to the text. Use the new style to unify the formatting of file names and paths throughout the documentation.

Using the custom :repo_link: directive, add a link to the ECM definitions folder for users that are interested in looking at additional ECM examples after completing the ECM authoring tutorial. Close #177.