Automate documentation build/preview
Add a Grunt task runner script to build a preview of the HTML version of the documentation and automatically rebuild the HTML when changes to the documentation files are detected. Add a section to the documentation outlining how to set up a computer to support Grunt and use the Grunt script.