Restrict lifetime advantage to lighting ECMs
The life cycle costing advantage given to ECMs with longer lifetimes has now been restricted to only lighting equipment ECMs. Without this restriction, ECMs with only a small lifetime advantage over a comparable baseline technology (e.g. 17 years vs. 16 years) would still see the benefit of an avoided baseline equipment purchase, even though such an advantage is probably in the noise of the ECM lifetime estimate. The advantage may be added back in for non-lighting ECMs in the future, given good reason to do so.

The sample ECMs used in the 'PrioritizationMetricsTest' and 'MetricUpdateTest' in 'run_test.py' were modified to be lighting equipment ECMs, such that the life cycle calculations for these ECMs are executed as they were before the above restriction was added.