Address #180
Fix issue #180. To make that change, update the calculator JavaScript to accommodate revisions to the formatting of the residential and commercial data, including slightly modified keys (e.g., 'electricity' instead of 'electricity (grid)' for that fuel type for residential buildings), the consolidation of the residential and commercial data into a single JSON file, and the addition of stock data for commercial buildings.

Reorganize and delete extraneous supporting files.

Update CO2 emissions notation from 'MMT' to 'Mt'.