Close #187
Correct commercial refrigeration data import arising from cases where names are not formatted correctly because they are so long that the scenario year (i.e., '2020' in '2020 typical') is partially truncated at the end of the string. Also fix greedy regex capture of extraneous spaces before the year in a variety of different technology description strings.

Correct the problem with residential lighting where data were not read because of a case mismatch for the incandescent lighting type, where in the 2015 AEO data, 'Inc' was used, and in the 2017 AEO data, 'INC' is used.