Remove residential ECM cost conversion from $/ft^2 floor to $/unit
The conversion relied on assumptions about the number of technology units per household that were difficult to verify. With this change, cost units of ‘$/ft^2 floor’ are sufficient for residential ‘add-on’ ECM types, though not for ‘full service’ residential ECMs, where a comparison between the $/unit cost of the ECM and a comparable baseline technology would be required.

Supporting cost conversion data in ‘ecm_cost_convert.json’ were removed; a few additional values in this file were modified to include additional significant figures; and envelope cost conversions units were changed from ‘null’ to explicitly state the conversion units.

Tests of the ‘$/ft^2 floor’ to ‘$/unit’ conversion in ‘ecm_prep_test.py’ were modified to reflect the removal of this conversion.