Slight adjustment to relative performance weighting
The relative energy performance level of baseline stock previously captured by an ECM is a weighted combination of the ECM’s relative performance level for the current year and all previous years it has been on the market.

Previously, the weights for this calculation were based on the proportions of competed and previously captured stock for the ECM. In the technical potential scenario, however, these proportions sometimes change dramatically from year to year, yielding sharp changes in the weighted relative performance value over time. With this change, a consistent baseline stock turnover rate is used for the weighting (1/baseline lifetime + retrofit rate).

A related update was made to one of the images in Tutorial 5 Figure 3.