Fix market share weights, heating/cooling overlap calculations
In ‘run.py,’ a previous commit had set a 100% weight for the market shares of competing ECMs in any year that one of the competing ECMs entered the market, for both the technical potential and max. adoption potential scenarios. This is now only the case for the technical potential scenario. In the maximum adoption scenario, a 100% weight is assigned when the competing ECM with the earliest market entry year enters the market; for all subsequent years, weights are based on the replacement lifetime of the existing stock in that market segment and the assumed retrofit rate.

A related correction was made to the baseline replacement lifetime data written out by ‘ecm_prep.py’ for each market segment.

Finally, the calculation that splits the savings impacts of supply-side and demand-side heating/cooling ECM is now robust to negative savings impacts.