Improved clarity of stock/flow dynamics in run.py
A more nuanced set of stock/flow calculations and associated explanations has been added to the measure competition routines in run.py. The updated calculations assume the following:

-In a technical potential case, ECMs are competed based on “long run” market shares that are adjusted in each year; these “long run” market shares assume 100% stock turnover in each year.
-In a maximum adoption potential case, ECMs compete for “new” and “existing” baseline technology stock differently:
*ECMs compete for all “new” stock that is added in each year as well as the portion of the “new” stock captured by ECMs in previous years that is up for replacement/retrofit.
*ECMs compete for the portion of “existing” baseline stock that is up for replacement/retrofit plus all “existing” stock previously captured by ECMs that is up for replacement/retrofit.
*The baseline stock turnover rate is dictated by the baseline technology’s lifetime and the assumed retrofit rate.
*The ECM stock turnover rate is dictated by the market share-weighted lifetime of all competing ECMs and the assumed retrofit rate.

ECM results in the /web folder were updated to reflect the above updates to the competition routine, which have a minimal effect on high-level savings outcomes.

In ecm_prep.py, an added check for comparable baseline technology lifetime data was implemented when cost and/or performance data are missing for the baseline technology (in some cases, baseline cost/performance data are zeroed out but lifetime data are still available).