Added ECM-specific retrofit rate and documentation
A user may now specify ECM-specific retrofit rates (‘retro_rate’) as part of ECM definitions; the rate may be specified as a point value or as a probability distribution. The latter required several enhancements to the ‘partition_microsegment’ function in ‘ecm_prep.py’.

The documentation was expanded to reflect the new ‘retro_rate’ parameter as part of the “Additional ECM features” section and to include both this parameter and the recently added ‘time_sensitive_valuation’ parameter in the ECM JSON schema.

Additionally, a bug was found in the stock-and-flow dynamics in ‘ecm_prep.py’ for secondary microsegments and was fixed, and a few other minor edits were made to the code.