Fixed problem determining number of structure types for envelope ECMs
Given an envelope ECM that applies to only one structure type, an overly broad if statement in ‘ecm_prep.py’ was always setting the ECM’s number of applicable structure types to two. This led to inaccuracies in the ECM’s total applicable baseline stock numbers and subsequent cost effectiveness calculations. The problematic if statement has been fixed, though a test of this ECM case should be added to ‘ecm_prep_test.py’ as described in issue #207.