Fix Travis flake8 error, update CI config
Update the CI config to use a newer version of numpy and to test against Python 3.7. Fix flake8 'W605 invalid escape sequence' errors by switching to raw strings for regexes and adding an escaping backslash to strings to be printed to stdout. W605 errors generally arise from regex definitions where there is a conflict between the use of the backslash character to denote special character classes, e.g., \s, and the use of the backslash character in string literals to denote other special functions, e.g., \n for newline.