Close #209
Update EIA data processing module test classes that check for raw EIA data files with an additional decorator that looks for an environment variable 'EXPECT_EIA_FILES' set to 'true'. If the environment variable is missing or set to another value, the tests that depend on the raw EIA data files are skipped.