Revised measure set, enabled sector shapes for FS measures, bug fix
The measure set in ./ecm_definitions was revised to incorporate updates to existing measures and to move existing measures that have not yet been updated or which will not be used in the decarbonization pathways analysis to another folder (./ecm_definitions/Other_ECMs).

Sector shape calculations for measures that switch from non-electric to electric equipment are now accommodated by the “gen_tsv_facts” function.

A bug in how EMM-resolved carbon intensity and cost data were keyed in as part of the analysis routines was fixed.