Improve stock and stock cost calculations
For equipment measures that apply to multiple end uses that include heating or cooling, stock and stock cost numbers are now anchored on a comparison to baseline heating technologies, if available, and baseline cooling technologies if not. This is particularly relevant to the consideration of heat pump measure stock and unit cost numbers in the context of building decarbonization analyses, where heating stock evolution is often a point of focus. Note that to support this approach, EIA’s baseline heat pump stock/unit cost numbers must be multiplied by 2 (since they are split across heating/cooling in EIA’s modeling).

For residential envelope measures, stock and stock costs are translated from a ft^2 floor basis to a home basis, where home is considered synonymous with “unit” for these measures thus allowing easier comparison and packaging with central HVAC equipment measures (where stock/stock costs are on a unit basis, and most homes are assumed to have only one central HVAC unit). Conversion factors are developed based on the total number of homes and number of total square feet that EIA projects for a given combination of region and building type. Any residential equipment measures where cost units are provided in $/ft^2 floor are now also translated to $/unit using the same approach as for residential envelope measures.

Whole building data needed to convert residential $/node or $/occupant inputs to $/ft^2 floor were removed from the ecm_cost_convert file under the assumption that cost data for such measures could be provided in $/ft^2 floor directly on the residential side.