Option to add BAU counterfactuals, suppress early replacements
The ‘--zero_ret’ and ‘add_typ_eff’ options suppress early equipment replacements and prepare BAU counterfactuals for all ESTAR/IECC/90.1-level equipment efficiency measures for further use in the competition analysis, respectively.

Note that BAU counterfactuals for equipment fuel switching measures are not handled by the ‘add_typ_eff’ option and must be manually specified with “Ref. Case” in the name (this substring is used to flag such measures in various parts of the code).

Furthermore, when ‘add_typ_eff’ is present with the existing ‘floor_start’ option, all measures will enter the market in the year specified in their JSON definition in ./ecm_definitions, and the BAU counterfactual measures will be removed from the market starting in the year provided alongside the ‘floor_start’ option.

Changed AEO start year to 2022 in anticipation of the upcoming calendar year change.