add --alt_regions_breakdown arguemnt for ecm_prep.py
Prior to this change if --alt_regions was used when calling ecm_prep.py
then an use was forced into specify the breakdown interactively.  This
prevented scripting a call to ecm_prep.py with --alt_regions.

This change introduces the --alt_regions_breakdown argument which takes
an interger value.  By default it will put the user into the interactive
specification, but also allows for the user to script the call and
breakdown.