Fix fuel splits under competition
An error in the application of fuel splits under competition in run.py has been fixed.

Special handling of stock turnover calculations in run.py for measures that apply to multiple end uses has been limited to equipment measures (envelope measures that apply to multiple end uses use square footage stock totals in a way that does not require this handling).

Additionally, the read-in of EMM data in run.py was adjusted to expect the new gzipped version of those data, and minor formatting errors were corrected.