Reduce size of written data files, refactor handling of user options
The data prepared by ecm_prep now exclude both competition and summary data for any ECMs that contribute to a package that is being prepared in the run; these data were extraneous because they are not actually used in run.py (only the package data is).

Handling of user options was refactored in ecm_prep and ecm_prep_test to streamline the use of the opts object as part of function inputs, and these options are now written out to a “usr_opts” attribute (formerly “energy_outputs”) for each ECM being prepared.