extend documentation re anaconda environment
Prior to this change the documentation noted that the conda environment
would provide the needed python and R software and dependencies.
However, there are conflicts in several libraries between R and python
making it difficult to have both in the conda environment.  Suggest that
end users use system level R and that we deprecate any use of R as soon
as possible.