Adjust definition of other heating measures
The definition of other heating measures (e.g., covering secondary heating and other heating equipment outside main classes) was adjusted to ensure complete separation between FS and non-FS applications, and to ensure that such measures would not be unintentionally removed from the market in certain cases (these measures had been assigned a ‘Ref. Case’ in their name, and all such measures are removed from the market when the –floor_start user option is present).

A small adjustment was made to stock/flow calculations to ensure that the annual competed stock fraction for stock being switched over to HPs is always between 0 and 1.