Corrections to envelope cost packaging
The calculations that support the –pkg_env_costs option were updated to fix an issue with how the envelope costs were being normalized to a per unit basis and added to the per unit costs of associated HVAC equipment in the package. In particular, the normalizing stock units in those calculations had not been properly harmonized across the envelope and HVAC data.

Additionally, a subset of “EE+DF” measures with envelope efficiency improvements was added to a new folder in ./ecm_definitions for easy future access. These special measures do not leverage the Scout packaging calculations in ecm_prep and therefore must reflect incremental envelope costs directly in their measure definitions to mimic the effects of using –pkg_env_costs in model runs.