Automate ref. case analogue prep for FS measures not using exogenous rates
Automated preparation of competing reference case analogues (via the ‘add_typ_eff’ user option) had excluded FS measures. This was oriented around the assessment of FS measures under exogenous FS rates, which precludes the need to compete the FS measures with their reference case, fossil-based analogues. However, when not assessing FS measures under exogenous FS rates, these reference case analogues should be prepared for direct competition with the FS measures – and they now are.

Additional user prompting was added to the ‘pkg_env_costs’ option to determine whether, when accounting for envelope costs in HVAC/envelope packages, the user also wants to prepare HVAC-only versions of the packages that can be subsequently competed with the full HVAC/envelope packages.

Finally, a bug in the preparation of the  attribute for some commercial measure cases was fixed.