Correction to ref. counterfactual prep, expansion of stock data reporting
In ecm_prep, for the endogenous FS case, reference case counterfactuals for the FS measure (e.g., fossil-based technologies) indicate None for the “fuel_switch_to” parameter, but this setting was inadvertently being applied to the FS measure that the counterfactual was developed based on; this issue was fixed.

In run, common stock turnover data used in measure competition calculations are now tied to uncompeted values, to decouple from any effects from the competition. Baseline stock values are now adjusted to reflect competition, and when the user executes the ‘--report_stk’ cmd line option, both uncompeted and post-competition baseline stock values are reported out, to facilitate an understanding of the effects of competition on measure baseline markets.