Add informative error messages
Users that try to run a measure requiring load savings data that aren’t present now receive an error message that indicates which file is missing and points them to the latest savings shape records on Zenodo.

Users that try to run a measure with ‘climate_zone’ inputs that are inconsistent with the high-level region setting for the run now receive an error message that instructs them to change either the ‘climate_zone’ or the higher-level region setting to be consistent.