Update ‘Best’ incremental envelope costs for Brattle runs
The ‘Best EE+DF’ tier of HVAC/envelope measures requires that incremental envelope costs be added directly to the measure definition when the intent is to include these costs in that of the package, since the Scout measure packaging feature is not used for this tier of measures (which are represented separately as a package in EnergyPlus).

The previous estimates of these incremental envelope costs were difficult to verify, thus, a new method for calculating them was devised and reflected in the costs (which are summarized here: https://docs.google.com/spreadsheets/d/1g09M5XISuHJus555G2J10bXpwprEyMMaJ1Imer9OnTQ/edit?usp=sharing).

The new method packages zero incremental cost ‘generic’ residential and commercial HVAC measures that cover all HVAC segments with ‘Best’ tier Scout envelope measures, and examines the cost differential between the efficient and baseline case to determine the incremental cost of the envelope portion of the package. These incremental envelope costs are then added to the ‘Best EE+DF’ HVAC equipment costs.