Handle all zero CSV baselines
Baselines in savings shape CSVs are all zero for building types that are inapplicable, but lack of handling of this resulted in problems dividing by zero when calculating efficient sector shapes. Add a check for zero baselines and if so, set the efficient shape to the baseline shape.

Also modify pool pumps measure to only apply to single family homes and add a small correction to one set of ‘Best’ ASHP measures that had not been updated to split performance out for secondary heating.