Measure changes, fix CPI, other small tweaks
The cost year was adjusted for several measures to ensure the definition reflects the cost units of primary data source(s) that inform the cost estimate. CAC and RTU efficiency measures/packages were removed from the measure set under the assumption that their potential is mostly displaced by HP adoption and they would not be a primary target of decarbonization policy; large commercial chiller efficiency measures remain, and prospective HPs are no longer assumed to apply to the large commercial segment. ‘Other’ residential electric heating conversion measures (e.g., secondary heating, wood stoves) are now separated out from HP packages into their own measures; they are also no longer separated by CC/non-CC, and no longer encompass GSHPs, which stay at reference levels.

The date column format in the CPI data was corrected to fix an error where improper format was resulting in no cost translation across years; a check for correct date format was also added.

For measure packages, the overall lifetime calculation was simplified to be the average of the lifetimes across all equipment measures in the package, excluding add-ons (unless the only measures in the package are add-ons.)

Measure competititon calculations are now on the basis of competed stock in a given year, vs. total stock.

Finally, all competition information for contributing microsegments that are prepared under exogenous HP switching rates is tagged with an additional ‘-FS’ (for fuel switching to HPs) or ‘-RST’ (for switching from resistance to HPs) to ensure that microsegment data for these measures is only further competed with microsegments for other HP measures that are prepared with the same exogenous rates. (A microsegment for an exogenous HP switching measure should not be competed with that of a gas- or cooling efficiency-only measure, since the exogenous switching rate already takes care of this competition.)