Improve cost year adjustments, update carbon and cost data
In ecm_prep.py, cost year translations are now based on averages of CPI values across an entire year (vs. the last values in each year).

In run.py, all calculations and reporting of costs are adjusted to a common year basis, where the common year is defined as the year before the current one (to ensure that a full year of CPI data exist).

Carbon price data were updated to reflect the latest White House publication on the social cost of carbon. CPI data updated to pull in last two months of data.

The gitignore file now ignores the file that would result from use of the cost optimization routine.