Fix missing cost year adjustment in ecm_prep for certain cases
The convention that sets commercial measure cost units in $/kBtu/h heating, cooling, or ventilation to those of the comparable baseline segments had precluded adjustments for cost year differences between measure and baseline cost units. This was fixed.

An error that was exposed in plots.py when running a measure set that only applies to certain regions (e.g., cold climates) was corrected.