Ensure no incentives for add-on measures
Add-on measures like controls generally pair with the equipment replacements that are incentivized, but do not receive those incentives in isolation, thus the code now screens for add-on cases when applying incentives. In the absence of this screen, measures like HVAC controls would receive full incentives for HVAC equipment replacements, which was not the intended behavior (only the equipment replacements should receive those incentives).