Fix prevention of adjustment to efficient fuel split data in run.py
For fuel switching measures, the code had restricted the competition adjustment in run.py of efficient case breakout data for the baseline fuel type to avoid unnecessary calculations and save computation time. This restriction assumes that exogenous switching rates are used  – by convention, efficient case breakout data for the baseline fuel type will be zero when using these rates and thus do not need any adjustment. When exogenous switching rates are not used, however, there will be efficient case breakout data for the baseline fuel type to adjust, and thus this restriction will lead to negative savings values for the baseline fuel type.

The fix checks for exogenous switching rates in this part of the run.py routine – when used, efficient case breakout data are assumed to be zero and not further adjusted, and when exogenous rates are not used, these breakout data are adjusted along with the base case data to yield the proper savings values.