Fix input metadata process, clean up file tests
Fix mseg_meta.py exception when importing files to extract the year range of the input data. The root cause of the exception was changes to converter.py in #2d6087f that require the user to specify the conversion file to be updated instead of using hardcoded file names. Instead, specify the conversion file path and add a test checking for the presence of the file.

Remove lingering references to Travis-CI in some test decorators and supporting test infrastructure.