Enable filtering of ECMs with new arguments
A new argument --ecm_files allows user to specify exact matches to ECM files that exist in the ECM directory. The --ecm_files_regex argument allows a list of regular expressions that match a subset of files in the ECM directory. One or both may be used to filter down the list of ECMs.