Continue ECM processing with errors, check all ECM inputs
Through try/except handling, errors preparing individual ECMs or ECM packages are now captured and written out to a text file in the ./generated folder. Skipped ECMs are also now registered to a new list (“skipped”) in ./generated/run_setup.json, and users are warned if there are any ECMs in this list when executing run.py. The update_active_measures() function is aware of and handles skipped measures. Fixes #255.

Additionally, pulled check_mkt_inputs() function out of fill_mkts() such that it is executed across all ECMs in the site before individual ECM markets are prepared. Fixes #369.

Modified the filter_invalid_packages() function to be more general, pulling the user warning out of the function and the corresponding test.