Adjust active and inactive measures based on user-selected ECMs
Because ECMs can be selected directly, there is a higher risk of unexpected measures getting run between Scout runs. This commit moves previously active measures to inactive if they are included in a package for the current ecm_prep step. Additionally, active measures will be moved to inactive if they are not within the user-selected set of ECMs, and a warning will be thrown indicating such.