Several minor fixes and enhancements
Remove secondary heating and wood stoves from being subject to exogenous switching rates, as they were not intended to be covered by these rates.

Allow users to set ‘tech_switch_to’ to the value ‘same’ and have the code calculate a switchover to the same technology under exogenous switching rates. This is sometimes necessary for accounting purposes – for example, if exogenous rates are represented and a baseline segment is included that is subject to those rates, but without a complementary measure in the analysis to switch that baseline segment to, the results will show a decline in that baseline segment without a complementary increase in a switched to segment, thus under-representing that baseline segment’s influence on the total results.

Add ASHP to technology list where zero costs are expected (in EIA new construction data ASHP costs are zero for cooling, all costs are in the heating end use).

Remove water services as a distinct technology type as it appears to be counted already under “unspecified” (see issue #394).

Refactor calculation of hourly TSV load adjustment factors to ensure that each end use in the tsv_load.gz data will be leveraged (previous approach had resulted in skipping some edge case end uses in those data like clothes washing, pool pumps and pool heaters that are under ‘other’ in the Scout data).

Ensure that once cumulative measure-captured or cumulative competed stock fractions reach 1 in a given modeling year, these fractions are not updated in any subsequent years of the modeling time horizon.

Improve warning message for units conversion issues.

Add measure JSON check to ensure that measures that are set to fuel switch to electricity apply to non-electric fuels in the baseline.

Suppress check for 'other' and 'unspecified' building type input breakouts when these categories are also present in measure end uses (prioritize correct end use input breakouts, which are more commonly used).