Link heating/cooling mseg turnover/switching calcs
Measures that apply to both heating and cooling equipment microsegments will now see stock turnover and (where applicable) exogenous HP switching rates linked across those microsegments. The linked calculations are anchored on a single end use and technology from the Measure’s applicable market. The default end use anchor is set to heating (in UsefulVars), while the anchor technology is determined uniquely for each Measure by comparing the technologies in the Measure’s ‘technology’ attribute to a candidate list of anchor technologies for each building type (res/com) and end use; the list order sets the priority for which anchor technology is used when the Measure applies to multiple candidate technologies in the list.

To ensure the anchor end use and technology are always updated first in the preparation routine, given the dependence of linked msegs on these updates, the iterable list of msegs to update for each Measure is reordered in fill_mkts so that msegs that include the anchor end use and technology come first.

Additional minor improvements in this commit:
* Add reporting of efficient stock that remains with baseline technology when applicable (previously only reported measure stock under the efficient key).
* Ensure that a measure's baseline and efficient market mseg breakouts are always adjusted to reflect measure competition when breakout values are not None or zero across all years.