Initial implementation of cooling additions (needs unit test)
Cooling additions for HP switching in homes without existing cooling are now handled in the partition_microsegment function. Calculation of these additions is triggered when the measure is flagged as switching to a heat pump (via ‘tech_switch_to’ parameter) and the cooling segment for the measure is scaled to zero via the ‘market_scaling_fractions’ parameter.

The implementation passes existing unit tests and has been run through successfully locally with sample measures outside of the test suite, but still requires a new sample measure in the unit tests.