Reintroduce coupled stock turnover in run
Commit 3f3d936 had erroneously removed the section of run.py that linked together stock turnover dynamics for cooling with heating msegs in the same measure (e.g., HPs), under the assumption that the new linkage between turnover for such msegs added in ecm_prep would cover this and be carried through to run.py via reporting of competed stocks for each of the linked msegs. In ecm_prep, however, competed cooling stock in such cases is ultimately set to zero to avoid double counting across those msegs; therefore, use of those competed cooling stock data in run.py made it appear as though there was no cooling stock turnover at all to drive the competition of cooling segments.

Adding back the linked turnover dynamics in run.py addresses this issue by ensuring such links rely on competed heating (or other linked mseg) stock data that are not set to zero in ecm_prep.

In reintroducing this code, a slight adjustment was needed to handle a new string that is now added to the measure competition information to signify the specific heating/cooling linkages that the measure represents, if any (see use of variable “tch_apnd”).

Finally, the rank order of heating/cooling technologies to choose as anchors for such within-measure linkages when multiple such technologies are present in the measure definition was harmonized across ecm_prep and run.